Removed old code, moved react code up a level.
