+++ bpaotu/bpaotu/frontend/src/containers/search_errors.tsx
