+++ bpaotu/bpaotu/frontend/src/containers/search_page.tsx
