+++ bpaotu/bpaotu/frontend/src/containers/search_results_table.tsx
