--- bpaotu/bpaotu/frontend/src/index.ts
@@ -1,7 +1,0 @@
-//import 'bootstrap';
-//import 'bootstrap/dist/css/bootstrap.min.css';
-
-// TODO revise if we should switch to multiple entry points instead.
-// ie. one entrypoint for each page
-export { init as init_tables } from './table_page';
-export { init as init_search } from './search_page';
