+++ bpaotu/bpaotu/frontend/src/interfaces.ts
@@ -3,10 +3,6 @@ import { SearchConfig } from './search';
 declare global {
   interface Window {
     otu_search_config: SearchConfig,
-    CKANAuthToken: string,
-    maps: any[],
-    loadmapsamples_map: any,
-    L: any
   }
 }
