--- bpaotu/bpaotu/frontend/src/print.ts
@@ -1,4 +1,0 @@
-
-export function info(msg: string) {
-    console.log(msg);
-}
