+++ bpaotu/bpaotu/frontend/src/reducers/search_page/utils.ts
