+++ bpaotu/bpaotu/frontend/tsconfig.json
@@ -6,9 +6,9 @@
     "lib": ["es2015", "es2017", "dom"],
     "jsx": "react",
     "allowJs": true
-  },
-  "files": [
-    "src/interfaces.ts",
-    "src/index.ts"
-  ]
+  }
+  //"files": [
+  //  "src/interfaces.ts",
+  //  "src/index.ts"
+  //]
 }
