+++ bpaotu/bpaotu/frontend/webpack.common.js
@@ -2,13 +2,7 @@ const path = require('path');
 const webpack = require('webpack');
 
 module.exports = {
-  entry: './src/react/init.tsx',
-  plugins: [
-    new webpack.ProvidePlugin({
-      $: 'jquery',
-      jQuery: 'jquery'
-    })
-  ],
+  entry: './src/init.tsx',
   module: {
     rules: [
       {
