--- bpaotu/bpaotu/static/common/map.css
@@ -1,33 +1,0 @@
-.leaflet-container {
-    /* all maps */
-    width: 800px;
-    height: 500px;
-}
-
-#site_map {
-    width: 100%;
-    height: 400px;
-}
-
-#specialbigmap {
-    height: 800px;
-}
-
-.leaflet-popup-content .site-popup-table {
-    max-height: 300px;
-    overflow-x: hide;
-    overflow-y: auto;
-}
-
-#sitelist .lat, #sitelist .lng {
-    display: none;
-}
-
-#sitelist-container {
-    overflow-y: auto;
-    overflow-x: hide;
-}
-
-h3.unstacked {
-    margin-top: 0px;
-}
